Machine Learning/api/requirements.txt
Flask
gunicorn
tensorflow
numpy
pillow
pandas
scikit-learn
mysql-connector-python
python-dotenv
